Code Reordering and improvement
