+++ tests/CMakeLists.txt
@@ -4,7 +4,7 @@ set(test_calcli_sources
 	test_cmap.cpp
 	test_operator.cpp
 	test_function.cpp
-	test_lexer.cpp
+	test_tokenization.cpp
 	test_infix_to_postfix.cpp
 	test_postfix_evaluation.cpp
 	test_calcli.cpp
